Refactored the handlers into packages
