+++ src/server/config/config.py
@@ -1,9 +1,9 @@
-from __main__ import (
+from ..app.app import (
     app,
     token_required,
 )
 
-import db
+from ..db import db
 from flask import request
 from flask_cors import cross_origin
 from sqlalchemy.exc import OperationalError
