+++ src/server/tests/run_tests.sh
@@ -1,19 +1,11 @@
 #!/bin/bash
-
 # POSTGRES
 eval `pifpaf run postgresql`
-
 # Run the tests
 cd ../..
-python3 -m server.tests.test
-
+python3 -m server.tests.test_login
 # Stop pifpaf
 pifpaf_stop
-
 # MYSQL
-
 # Run the tests
-
 # Stop pifpaf
-
-
